Move private stuff into `impl.ts`
